Removed Right hand rotation control on crosshair
Here I just opted to remove the Right Hand Rotation to be in sync with crosshairs, cause the jittering was just really annoying when in close proximity to walls and such.
for instance, checking out fortnite footage, I came to the conclusion that in fortnite they do the exact same thing, without syncing the barrel orientation with crosshairs.