BugFix, when swapping weapon & aiming
Here I fixed the fact that we could reload when swapping weapons,
also the sniper and M4 scope wouldn't be hidden
and if the character is aiming, will now stop aiming upon swapping weapons.

Note: there is still a bug where the you can't see the sniper mesh or M4 mesh after the bool forcefully pushes you out of ADS, this is because we're hiding the weaponmesh upon aiming with those.
will fix later with the Enum add to combat states: ECS_Swapping.