Important implementations (player joined/quit game scoreboard logic), and others
now when players join or quit the game, the scoreboard will update the stat line and remove it, also the local player will always now have its own stat line green, to see itself more easily when tabbing, also a portal in the lobby was introduced to make it possible to travel straight into the game.
fixed a bug where Berserker was making the gun not shoot after the timer.
Menu changes, now the settings menu is inside the return to main menu menu.
the Quit button is now in C++ and it will execute all of the correct logic in order to remove the player stat line from the scoreboard and also execute the elim() to drop the weapon