Voting System Bugfix (servertravel and menu teardown)
Fixed the last 2 hitches with the voting system.
The first issue was that servertravel was completely destroying the HUD, which wasn't showing up once the map was changed.
And the other issue was that upon calling MenuTeardown for the voting system, it was only being called on the server, so now I set it as a timer, so that it gets destroyed shortly before the servertravel happens.
Also the variable has been exposed to BPs, so it now can be adjusted accordingly, depending on the length of the cooldown timer.