Implementation of Icons Lightup on buff pickup (jump, speed and Berserk)
Now we finally have a visual representation of the buffs.
An icon will now light up properly for the local player that picks up one of these buffs, giving the player an idea of what's happening to the buffs he just picked up, which won't leave it up to guessing.