Implementation Starter menu rework, Server List.
Now it's possible to hit the server search button and look for all available servers, and join whichever the player wants.
NB: It's done by writing the whole serverID in the editable text box, which is not ideal //TODO find a solution for it.
Code is also a little messy, requires a little cleanup.